fix(functional-tests): register WAF bypass header on the browser context
Because:
- addWafBypassHeader used page.route(), a per-page handler, so pages opened
  outside the fixtures did not inherit it. ConfigPage.getConfig() opens a
  throwaway page via context().newPage(); on stage/production the content
  server sits behind Fastly's bot "Client Challenge", so getConfig() timed
  out waiting for meta[name="fxa-config"] and every beforeEach that calls it
  failed (cms, passkey, and related suites).

This commit:
- Registers the WAF bypass route on the browser context instead of the page,
  so every page in the context (including getConfig's newPage) inherits it,
  while keeping the FXA-domain scoping that avoids leaking the token to
  third-party origins.
- Extracts addWafBypassHeader into lib/waf.ts to avoid a pages<->fixtures
  import cycle, re-exported from fixtures/standard for existing importers.
