+++ packages/functional-tests/lib/fixtures/standard.ts
@@ -17,8 +17,11 @@ import { BaseTarget } from '../targets/base';
 import { TestAccountTracker } from '../testAccountTracker';
 import { PasskeyPage } from '../../pages/passkey';
 import { GleanEventsHelper } from '../glean';
+import { addWafBypassHeader } from '../waf';
 import { existsSync, readFileSync } from 'fs';
 import { join, dirname, basename } from 'path';
+
+export { addWafBypassHeader };
 
 // The DEBUG env is used to debug without the playwright inspector, like in vscode
 // see .vscode/launch.json
@@ -34,42 +37,6 @@ export type TestOptions = {
   gleanEventsHelper: GleanEventsHelper;
 };
 export type WorkerOptions = { targetName: TargetName; target: ServerTarget };
-
-const CI_WAF_TOKEN = process.env.CI_WAF_TOKEN;
-
-/**
- * Adds a route handler that injects the WAF bypass header on requests to
- * FXA-owned domains only, leaving third-party origins (Stripe, hCaptcha)
- * untouched. No-op when CI_WAF_TOKEN is unset.
- */
-export async function addWafBypassHeader(page: Page, target: BaseTarget) {
-  if (!CI_WAF_TOKEN) {
-    if (target.name === 'stage' || target.name === 'production') {
-      // eslint-disable-next-line no-console
-      console.warn(
-        `⚠ CI_WAF_TOKEN is not set for target "${target.name}". Requests may be blocked by the WAF.`
-      );
-    }
-    return;
-  }
-  const fxaDomains = [
-    new URL(target.contentServerUrl).host,
-    new URL(target.authServerUrl).host,
-    new URL(target.paymentsNextUrl).host,
-    new URL(target.relierUrl).host,
-  ];
-  const pattern = new RegExp(
-    fxaDomains.map((d) => d.replace(/[.*+?^${}()|[\]\\]/g, '\\$&')).join('|')
-  );
-  await page.route(pattern, async (route) => {
-    await route.continue({
-      headers: {
-        ...route.request().headers(),
-        'fxa-ci': CI_WAF_TOKEN,
-      },
-    });
-  });
-}
 
 export const test = base.extend<TestOptions, WorkerOptions>({
   targetName: ['local', { scope: 'worker', option: true }],
